seo: edit page (settings)
